requirements.txt
schedule~=1.2.2
pytelegrambotapi~=4.22.1
python-gitlab~=4.9.0
Flask~=3.0.3
PyYAML~=6.0.2
requests~=2.32.3